% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ncomb.R
\name{Ncomb}
\alias{Ncomb}
\title{Variable Information Dataframe}
\usage{
Ncomb(.data)
}
\arguments{
\item{.data}{The name of the dataset to summarize.}
}
\value{
A dataframe
}
\description{
This function takes the variable name and type and combines into a singular dataframe with the summary statistics.
}
\examples{
Ncomb(iris)
}
