% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFun.R
\name{NFun}
\alias{NFun}
\title{Summary Statistics}
\usage{
NFun(.data)
}
\arguments{
\item{.data}{The name of the dataset to summarize.}
}
\value{
A data frame
}
\description{
This function will grab the summary statistics from a dataframe including the number of observations, if the variable/column has any missing values, as well as the minimum and maximum for numeric data.
}
\examples{
NFun(iris)

}
