% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfout.R
\name{pdfout}
\alias{pdfout}
\title{Read Table/Codebook out to PDF File}
\usage{
pdfout(.data, pdftitle)
}
\arguments{
\item{.data}{The name of the dataset to summarize.}

\item{pdftitle}{The pdf file name that will be saved.}
}
\value{

}
\description{
This function outputs a PDF containing a table of the variable names, types and summary statistics.
}
\examples{
pdfout(iris, "IrisCodebook2020")
}
