% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyTable.R
\name{ShinyTable}
\alias{ShinyTable}
\title{Variable Summaries Shiny Gadget}
\usage{
ShinyTable(.data)
}
\arguments{
\item{.data}{The name of the dataset to summarize.}
}
\value{

}
\description{
This shiny gadget allows you to choose one variable at a time and then displays the summary information about that variable.
}
\examples{
ShinyTable(iris)
}
