% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablfun.R
\name{tablfun}
\alias{tablfun}
\title{Formatted Table of Codebook Information}
\usage{
tablfun(
  .data,
  headcolor = "white",
  bodycolor = "white",
  bodftcolor = "black",
  topftcolor = "black"
)
}
\arguments{
\item{.data}{The name of the dataset to summarize.}

\item{headcolor}{The background color of the header of the table.}

\item{bodycolor}{The background color of the body of the table.}

\item{bodftcolor}{The font color of the header of the table.}

\item{topftcolor}{The font color of the body of the table.}
}
\value{

}
\description{
This function takes the combined dataframe containing the information about the dataset and uses flextable to make it into a clean, formatted table that can be used as a codebook within the R markdown file or output to a PDF to share with collaborators.
}
\examples{
tablfun(iris, "pink", "yellow", "black", "blue")
}
