% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varinfo.R
\name{varinfo}
\alias{varinfo}
\title{Variable Names/Types}
\usage{
varinfo(.data)
}
\arguments{
\item{.data}{The name of the dataset to summarize.}
}
\value{

}
\description{
This function will grab the variables names from the chosen input data set and list them in a dataframe alongside their class.
}
\examples{
varinfo(iris)

}
